moved routes to the new and old structure respectively
