+++ newstructure/new_routes.go
@@ -1,4 +1,4 @@
-package main
+package newstructure
 
 import (
 	"fmt"
